Add files via upload
The transmission line is the most crucial part of the power system.
the prominent role of a transmission line is to transmit electric power from the
source area to the distribution network. The electrical power system consists of
so many complex dynamic and interacting elements that are always prone to disturbance or an electrical fault.